# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EdgeEver API",
    "version": "0.1.0",
    "description": "REST API for EdgeEver notebooks, memos, resources, API tokens, and Agent integrations."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/memos/{id}/share": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MemoId"
        }
      ],
      "get": {
        "summary": "Read the active public share for a memo",
        "responses": {
          "200": {
            "description": "Current share status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "share": {
                      "oneOf": [
                        { "$ref": "#/components/schemas/MemoShare" },
                        { "type": "null" }
                      ]
                    }
                  },
                  "required": ["share"]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or return a public read-only share link",
        "responses": {
          "200": {
            "description": "Existing share",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "share": { "$ref": "#/components/schemas/MemoShare" } },
                  "required": ["share"]
                }
              }
            }
          },
          "201": {
            "description": "Share created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "share": { "$ref": "#/components/schemas/MemoShare" } },
                  "required": ["share"]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Revoke a public memo share",
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/public/shares/{token}": {
      "parameters": [
        {
          "name": "token",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "summary": "Read a publicly shared note",
        "security": [],
        "responses": {
          "200": {
            "description": "Public read-only note",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "share": { "$ref": "#/components/schemas/PublicMemoShare" } },
                  "required": ["share"]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/public/shares/{token}/resources/{resourceId}/blob": {
      "parameters": [
        {
          "name": "token",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        },
        {
          "name": "resourceId",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "summary": "Read an attachment belonging to a publicly shared note",
        "security": [],
        "responses": {
          "200": { "description": "Shared resource blob" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "edgeever_session"
      }
    },
    "parameters": {
...
      "MemoId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Ok": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Ok"
            }
          }
        }
      },
      "Error": {
        "description": "API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": ["ok"]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "MemoShare": {
        "type": "object",
        "properties": {
          "memoId": { "type": "string" },
          "token": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["memoId", "token", "createdAt", "updatedAt"]
      },
      "PublicMemoShare": {
        "type": "object",
        "properties": {
          "title": { "type": ["string", "null"] },
          "contentJson": { "$ref": "#/components/schemas/TiptapDoc" },
          "contentMarkdown": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["title", "contentJson", "contentMarkdown", "tags", "updatedAt"]
      },
...
      "TiptapDoc": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string",
            "const": "doc"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": ["type", "content"]
      },
...
    }
  }
}